+++ ng/str.c
@@ -1,0 +1,4 @@
+
+/* Copyright (C) 2022 - 2023 Nathanael Specht */
+
+